Refactor UsersController and update database schema
- Updated UsersController to handle user ID as a route parameter for PUT and GET requests.
- Added Microsoft.EntityFrameworkCore.Tools package for migrations.
- Modified Program.cs to register new services and activate AppDbContext.
- Updated FinanceTracker.Data.csproj with Microsoft.EntityFrameworkCore.Relational reference.
- Improved phone number validation in LocalPhoneAttribute.
- Changed UserId and CategoryId types from int to long in Budget, Category, and Transaction classes.
- Annotated User class with [Table] attribute for database schema.
- Added AutoMapper.Extensions.Microsoft.DependencyInjection reference in FinanceTracker.Service.csproj.
- Created ServiceExtension class for service registration.
- Added migrations to set up Users, Budgets, Categories, and Transactions tables.
- Updated AppDbContextModelSnapshot to reflect current database model.
- Renamed Users table to users under auth schema and dropped unnecessary foreign keys in the second migration.

+++ src/FinanceTracker.Domain/Entities/Transaction.cs
@@ -5,9 +5,9 @@ namespace FinanceTracker.Domain.Entities;
 
 public class Transaction : Auditable
 {
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User User { get; set; }
-    public int CategoryId { get; set; }
+    public long CategoryId { get; set; }
     public Category Category { get; set; }
     public decimal Amount { get; set; }
     public TypeEnums Type { get; set; }
